Convert tasks to new function-based handler style
