+++ packages/shared-src/src/constants/jobs.js
@@ -4,7 +4,7 @@ export const JOB_TOPICS = {
       ALL_FROM_UPSTREAM: 'fhir.refresh.allFromUpstream',
       ENTIRE_RESOURCE: 'fhir.refresh.entireResource',
       FROM_UPSTREAM: 'fhir.refresh.fromUpstream',
-      RESOLVER: 'fhir.refresh.resolver',
     },
+    RESOLVER: 'fhir.resolver',
   },
 };
