--- packages/sync-server/app/tasks/FhirRefreshResolver.js
@@ -1,8 +1,0 @@
-import { WorkerTask } from 'shared/tasks';
-
-export class FhirRefreshResolver extends WorkerTask {
-  async doWork() {
-    this.log.debug('Running FHIR upstream resolver');
-    await this.sequelize.query('CALL fhir.resolve_upstreams()');
-  }
-}
