+++ packages/sync-server/app/tasks/fhir/refresh/fromUpstream.js
@@ -1,0 +1,32 @@
+import { FHIR_INTERACTIONS, JOB_SUBMITS } from 'shared/constants';
+import { resourcesThatCanDo } from 'shared/utils/fhir/resources';
+
+const materialisableResources = resourcesThatCanDo(FHIR_INTERACTIONS.INTERNAL.MATERIALISE);
+
+export async function fromUpstream(
+  { payload: { resource, upstreamId } },
+  { log, models: { FhirJob } },
+) {
+  log.debug('Finding resource by name', { resource });
+  const Resource = materialisableResources.find(({ fhirName }) => fhirName === resource);
+  if (!Resource) {
+    throw new Error(`FhirRefreshFromUpstream: No materialisable resource found for ${resource}`);
+  }
+
+  log.debug('Starting materialise', { resource, upstreamId });
+  const result = await Resource.materialiseFromUpstream(upstreamId);
+  log.debug('Done materialising', {
+    resource,
+    upstreamId,
+    resourceId: result.id,
+    versionId: result.versionId,
+  });
+
+  await FhirJob.submit(
+    JOB_SUBMITS.FHIR.RESOLVER,
+    {},
+    {
+      discriminant: 'one at a time',
+    },
+  );
+}
