+++ packages/sync-server/app/tasks/fhir/resolver.js
@@ -1,0 +1,4 @@
+export function resolver({}, { log, sequelize }) {
+  log.debug('Running FHIR upstream resolver');
+  return sequelize.query('CALL fhir.resolve_upstreams()');
+}
